refactor: add NavbarItem and Stories
