+++ packages/gatsby-theme-project-portal/src/components/Navbar.tsx
@@ -3,38 +3,7 @@ import { Link } from "gatsby"
 import { FaBars, FaTimes } from "react-icons/fa"
 import { IGatsbyImageData } from "gatsby-plugin-image"
 import { SiteTitle } from "./SiteTitle"
-
-interface NavbarItemProps {
-  name: string
-  link?: string
-  isActive: boolean
-}
-
-const NavbarItem: FunctionComponent<NavbarItemProps> = ({
-  name,
-  link,
-  isActive,
-}) => {
-  return (
-    <>
-      <li className="nav-item">
-        <span className="flex items-center p-5 leading-snug text-white hover:opacity-75 xl:text-black xl:px-3 xl:py-2">
-          <Link to={link ? link : "#pablo"}>
-            {isActive ? (
-              <span className="text-nav pb-1 ml-2 font-bold border-b-2 border-white xl:border-primary">
-                {name}
-              </span>
-            ) : (
-              <span className="text-nav pb-1 ml-2 border-b-2 hover:border-primary xl:border-transparent">
-                {name}
-              </span>
-            )}
-          </Link>
-        </span>
-      </li>
-    </>
-  )
-}
+import { NavbarItem } from "./NavbarItem"
 
 interface NavbarLayoutProps {
   title: string
